fixed two warnings
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@18871 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/kits/media/DefaultMediaTheme.cpp
@@ -995,7 +995,7 @@ DefaultMediaTheme::MakeViewFor(BParameter *parameter, const BRect *hintRect)
 				// ToDo: take BContinuousParameter::GetResponse() & ValueStep() into account!
 
 				for (int32 i = 0; i < continuous.CountChannels(); i++)
-					slider->SetLimitsFor(i, continuous.MinValue() * 1000, continuous.MaxValue() * 1000);
+					slider->SetLimitsFor(i, int32(continuous.MinValue() * 1000), int32(continuous.MaxValue() * 1000));
 
 				return slider;
 			}
